requirements-prod.txt
# TokenDNA — production-only dependencies
#
# Use this in production deploys instead of (or in addition to) requirements.txt:
#
#   pip install -r requirements.txt -r requirements-prod.txt
#
# These deps are not needed for tests / dev — they are runtime-optional and
# behave as no-ops when missing (see modules/observability/* and the secret
# gate).

# ── Database migrations ──────────────────────────────────────────────────────
alembic==1.13.3
psycopg-pool==3.2.4

# ── Observability (opt-in via env vars) ──────────────────────────────────────
prometheus-client==0.21.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
sentry-sdk==2.18.0
